requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.6
dnspython==2.1.0
idna==3.2
pymongo==3.12.0
python-telegram-bot==13.7
pytz==2021.3
requests==2.26.0
six==1.16.0
tornado==6.1
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
